Make encoder return a long representing (length, offset) and update API to compile against this

+++ fix-gateway-codecs/src/main/java/uk/co/real_logic/fix_gateway/builder/OrderSingleEncoder.java
@@ -41,7 +41,7 @@ public class OrderSingleEncoder implements MessageEncoder
     {
     }
 
-    public int encode(final MutableAsciiBuffer buffer, final int offset)
+    public long encode(final MutableAsciiBuffer buffer, final int offset)
     {
         return 0;
     }
